DP- The robot coin problem using obstacles
The following program is the popular robot coin problem demonstrating the use of dynamic programming. However in this case, the robot is given obstacles and must find the optimal path to achieve the maximum amount of coins possible. This program uses a bottom up approach in which the solution is shown in the bottom right corner.